chyrp-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
sqlalchemy
python-multipart
pydantic
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
google-generativeai
redis
fastapi-cache2[redis]
asyncpg
requests
